requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
langchain-cohere
faiss-cpu
wikipedia
langchain
sentence_transformers
langchainhub
langchain-groq
beautifulsoup4
langchain_groq
flask